pipeline runs through ccf for both yeast and human; test profiles exist for both; genome from aws; igenomes not working
