--- assets/GCF_000146045.2_R64_genomic.fa
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:fe42735d3e6242f8c40b457a232940e34ef15eb4502717e54a34ed118face4c1
-size 12310392
