--- assets/hg38_chr1_subset.fa
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a370fdbc088a5f243c8dd4f682e62421a755ff84e0bc28d4e13716064cead826
-size 91125010
