+++ assets/human_input_samplesheet.csv
@@ -1,0 +1,2 @@
+sample,fastq_1,fastq_2,barcode_details
+human_small,human/AY53-1_50k_downsampled_human.fastq.gz,,mammals_barcode_details.json
