--- assets/input_samplesheet.csv
@@ -1,2 +1,0 @@
-sample,fastq_1,fastq_2,barcodes
-human_small,TEST_INPUT.fq.gz,,TEST_INPUT.txt
