--- assets/mammals_barcode_components_indicies.json
@@ -1,6 +1,0 @@
-{
-    "om_pb": [0, 3],
-    "pb_lrt1": [3, 28],
-    "srt": [28, 32],
-    "pb_lrt2": [32, 38]
-}
