+++ conf/test_human.config
@@ -15,8 +15,9 @@ params {
     config_profile_name        = 'Test profile'
     config_profile_description = 'Minimal test dataset to check pipeline function'
 
-    input                      = "${projectDir}/assets/input_samplesheet.csv"
-    fasta                      = "${projectDir}/assets/hg38_chr1_subset.fa"
+    input                      = "${projectDir}/assets/human_input_samplesheet.csv"
+    organism                   = 'mammal'
+    fasta                      = "${params.igenomes_base}/Homo_sapiens/NCBI/GRCh38/Sequence/WholeGenomeFasta/genome.fa"
     aligner                    = 'bwamem2'
     r1_bc_pattern              = 'NNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNN'
     barcode_length             =  38
