+++ conf/test_yeast.config
@@ -16,14 +16,15 @@ params {
     config_profile_description = 'Minimal test dataset to check pipeline function for yeast'
 
     input                      = "${projectDir}/assets/yeast_input_samplesheet.csv"
-    fasta                      = "${projectDir}/assets/yeast_chr1.fa"
+    organism                   = 'yeast'
+    fasta                      = "${params.igenomes_base}/Saccharomyces_cerevisiae/Ensembl/R64-1-1/Sequence/WholeGenomeFasta/genome.fa"
     aligner                    = 'bwamem2'
     r1_bc_pattern              = 'NNNNNNNNNNNNNNNNNNNNNN'
     r2_bc_pattern              = 'NNNNNNNNNNNN'
     barcode_length             =  34
     min_mapq                   =  10
-    promoter_bed               = "${projectDir}/assets/yeast_test_promoters.bed"
-    background_data            = "${projectDir}/assets/yeast_background_data.csv"
+    promoter_bed               = "${projectDir}/assets/rmlab_notOrf.bed"
+    background_data            = "${projectDir}/assets/yeast_S288C_dSir4_background.bed"
     pileup_stranded            = 'false'
     insertion_length           = 1
     max_cpus                   = 10
