+++ nextflow.config
@@ -11,6 +11,7 @@ params {
 
     // Input options
     input                      = null
+    organism                   = null
     promoter_bed               = null
     background_data            = null
     pileup_stranded            = null
@@ -49,6 +50,7 @@ params {
     // Boilerplate options
     outdir                     = './results'
     tracedir                   = "${params.outdir}/pipeline_info"
+    publish_dir_mode           = 'copy'
     email                      = null
     email_on_fail              = null
     plaintext_email            = false
