+++ nextflow_schema.json
@@ -12,14 +12,15 @@
             "description": "Define where the pipeline should find input data and save output data.",
             "required": [
                 "input",
+                "organism",
                 "aligner",
+                "min_mapq",
                 "barcode_length",
                 "insertion_length",
                 "r1_bc_pattern",
                 "promoter_bed",
                 "background_data",
-                "pileup_stranded",
-                "min_mapq"
+                "pileup_stranded"
             ],
             "properties": {
                 "input": {
@@ -31,6 +32,10 @@
                     "description": "Path to comma-separated file containing information about the samples in the experiment.",
                     "help_text": "You will need to create a design file with information about the samples in your experiment before running the pipeline. Use this parameter to specify its location. It has to be a comma-separated file with 3 columns, and a header row. See [usage docs](https://nf-co.re/callingcards/usage#samplesheet-input).",
                     "fa_icon": "fas fa-file-csv"
+                },
+                "organism": {
+                    "type": "string",
+                    "default": "None"
                 },
                 "aligner": {
                     "type": "string",
@@ -278,6 +283,10 @@
                     "default": "${params.outdir}/pipeline_info",
                     "fa_icon": "fas fa-cogs",
                     "hidden": true
+                },
+                "publish_dir_mode": {
+                    "type": "string",
+                    "default": "copy"
                 },
                 "validate_params": {
                     "type": "boolean",
